+++ src/components/common/SubscribeButton.tsx
@@ -1,17 +1,39 @@
 import { Button, ButtonProps } from "@mui/material";
-import { MouseEvent } from "react";
+import { MouseEvent, useEffect, useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { RootState } from "../../state/store.ts";
+import { subscribe, unSubscribe } from "../../state/features/videoSlice.ts";
 
 interface SubscribeButtonProps extends ButtonProps {
   name: string;
 }
 
-const isSubscribed = false;
 export const SubscribeButton = ({ name, ...props }: SubscribeButtonProps) => {
+  const dispatch = useDispatch();
+  const select = useSelector((state: RootState) => {
+    return state.video;
+  });
+  const [isSubscribed, setIsSubscribed] = useState<boolean>(false);
+
+  useEffect(() => {
+    setIsSubscribed(select.subscriptionList.includes(name));
+  }, []);
+
+  const subscribeToRedux = () => {
+    dispatch(subscribe(name));
+    setIsSubscribed(true);
+  };
+  const unSubscribeFromRedux = () => {
+    dispatch(unSubscribe(name));
+    setIsSubscribed(false);
+  };
+
   const manageSubscription = (e: MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
     e.stopPropagation();
-    console.log("subscribed to: ", name);
+    isSubscribed ? unSubscribeFromRedux() : subscribeToRedux();
   };
+
   const verticalPadding = "3px";
   const horizontalPadding = "8px";
   const buttonStyle = {
@@ -23,7 +45,7 @@ export const SubscribeButton = ({ name, ...props }: SubscribeButtonProps) => {
     paddingLeft: horizontalPadding,
     paddingRight: horizontalPadding,
     borderRadius: 28,
-    display: "none",
+    height: "40px",
   };
   return (
     <Button
