+++ src/constants/Misc.ts
@@ -1,4 +1,6 @@
 export const minPriceSuperlike = 10;
 export const titleFormatter = /[^a-zA-Z0-9\s-_!?()&'",.;:|—~@#$%^*+=<>]/g;
+export const titleFormatterOnSave = /[^a-zA-Z0-9\s-_!()&',.;—~@#$%^+=]/g;
 
-export const titleSaveFormatter = /[^a-zA-Z0-9\s-_!()&',.;—~@#$%^+=]/g;
+export const allTabValue = "all";
+export const subscriptionTabValue = "subscriptions";
