+++ src/wrappers/GlobalWrapper.tsx
@@ -15,7 +15,7 @@ import {
   setSuperlikesAll,
   setUserAvatarHash,
 } from "../state/features/globalSlice";
-import { VideoPlayerGlobal } from "../components/common/VideoPlayerGlobal";
+import { VideoPlayerGlobal } from "../components/common/VideoPlayer/VideoPlayerGlobal.tsx";
 import { Rnd } from "react-rnd";
 import { RequestQueue } from "../utils/queue";
 import { EditVideo } from "../components/EditVideo/EditVideo";
